pyproject.toml
[tool.poetry]
name = "fasttrace"
version = "0.1.0"
description = ""
authors = ["Sakti Dwi Cahyono <sakti@sampingan.co.id>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.2"
opentelemetry-instrumentation-fastapi = "^0.24b0"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
opentelemetry-sdk = "^1.5.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.5.0"
sqlmodel = "^0.0.4"
opentelemetry-instrumentation-sqlalchemy = "^0.24b0"
aiosqlite = "^0.17.0"
asyncpg = "^0.24.0"
databases = "^0.5.2"
opentelemetry-instrumentation-asyncpg = "^0.24b0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^21.9b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
